+++ shield_plugin/example/android/app/src/main/kotlin/com/example/plugin_shieldfraud_example/MainActivity.kt
@@ -1,0 +1,7 @@
+package com.example.plugin_shieldfraud_example
+
+import io.flutter.embedding.android.FlutterActivity
+import io.flutter.embedding.android.FlutterFragmentActivity
+
+class MainActivity : FlutterActivity() {
+}